refactor: useThread 훅 분리
